stopped wall sticking
I added a custom material to be added onto the capule collider of the player so that he doenst get stuck when bumping next to a wall. this will allow the player to run at same speed as normal and bumping into the steps will now not affect the jump.