added Chase Range
I added a circular radius around the enemy that detects if player is withen range and only then does the enemy move towards the player. this will be useful down the road when I have the actual map ready and there are multiple rooms with more enemies. It will make it less hectic because less zombies chasing player but also it will make it more realistic as it will simulate that zombies are no longer aware of where you went and go back into their hibernation state.