Enemy Movement& Gun
I have developed the enemys movement some more, it now stops a metre away from the player and outputs a phrase in the console for now as I have not coded in the attack yet. I have also added a gun model I found on the Unity Asset Store and have attached it to foolow the camera.