Multiple things done
I have made a hit effect on the wall when the player hits the wall. In the future I could add diffrent hit effects for diffrent materials, but for now this will do.
I have animated the prototype enemys attack and idle states, and now does damage to the player, I have also added a rotation to the enemy so now the enemey can rotate to face the player at when in it is chasing or attacking the player.