flashlight
I added a flashlight for the player to use in the game. I added a decay effect for the torch to lose battery over time which the player can refill by finding batteries across the map.  I still have to make the flashlight pickup batteries. but why do something today that I can do tomorrow?