Light Decay and Battery
I created a battery script and a light decay so that when the layer walks through the battery his torch gains energy. I will add NV next and have it so you can switch between them with the left hand like how you can switch guns with the right hand.